wsclient/requirements.txt
python-socketio==5.3.0
requests==2.27.1